+++ packages/sdk/test/OperationRequestBuilder.test.ts
@@ -34,15 +34,18 @@ describe("OperationRequestBuilder", () => {
         },
       ],
       chainId: 1n,
+      tellerContract: DUMMY_CONTRACT_ADDR,
       deadline: 2n,
     };
 
-    const builder = new OperationRequestBuilder();
+    const builder = new OperationRequestBuilder({
+      chainId: 1n,
+      tellerContract: DUMMY_CONTRACT_ADDR,
+    });
     const opRequest = builder
       .action(DUMMY_CONTRACT_ADDR, getDummyHex(0))
       .unwrap(shitcoin, 3n)
       .refundAsset(shitcoin)
-      .chainId(1n)
       .deadline(2n)
       .build();
 
@@ -73,16 +76,19 @@ describe("OperationRequestBuilder", () => {
         },
       ],
       chainId: 1n,
+      tellerContract: DUMMY_CONTRACT_ADDR,
       deadline: 2n,
     };
 
-    const builder = new OperationRequestBuilder();
+    const builder = new OperationRequestBuilder({
+      chainId: 1n,
+      tellerContract: DUMMY_CONTRACT_ADDR,
+    });
     const opRequest = builder
       .action(DUMMY_CONTRACT_ADDR, getDummyHex(0))
       .unwrap(shitcoin, 3n)
       .refundAsset(shitcoin)
       .confidentialPayment(shitcoin, 1n, receiver)
-      .chainId(1n)
       .deadline(2n)
       .build();
 
@@ -113,10 +119,14 @@ describe("OperationRequestBuilder", () => {
       gasPrice: 30n,
       maxNumRefunds: 1n,
       chainId: 1n,
+      tellerContract: DUMMY_CONTRACT_ADDR,
       deadline: 2n,
     };
 
-    const builder = new OperationRequestBuilder();
+    const builder = new OperationRequestBuilder({
+      chainId: 1n,
+      tellerContract: DUMMY_CONTRACT_ADDR,
+    });
     const opRequest = builder
       .action(DUMMY_CONTRACT_ADDR, getDummyHex(0))
       .unwrap(shitcoin, 3n)
@@ -127,7 +137,6 @@ describe("OperationRequestBuilder", () => {
         gasPrice: 30n,
       })
       .maxNumRefunds(1n)
-      .chainId(1n)
       .deadline(2n)
       .build();
 
@@ -162,14 +171,17 @@ describe("OperationRequestBuilder", () => {
       refundAssets: [],
       actions: [],
       chainId: 1n,
+      tellerContract: DUMMY_CONTRACT_ADDR,
       deadline: 2n,
     };
 
-    const builder = new OperationRequestBuilder();
+    const builder = new OperationRequestBuilder({
+      chainId: 1n,
+      tellerContract: DUMMY_CONTRACT_ADDR,
+    });
     const opRequest = builder
       .confidentialPayment(shitcoin, 1n, receivers[0])
       .confidentialPayment(stablescam, 2n, receivers[1])
-      .chainId(1n)
       .deadline(2n)
       .build();
 
@@ -237,10 +249,14 @@ describe("OperationRequestBuilder", () => {
       refundAddr: refundAddr,
       actions,
       chainId: 1n,
+      tellerContract: DUMMY_CONTRACT_ADDR,
       deadline: 2n,
     };
 
-    const builder = new OperationRequestBuilder();
+    const builder = new OperationRequestBuilder({
+      chainId: 1n,
+      tellerContract: DUMMY_CONTRACT_ADDR,
+    });
     const opRequest = builder
       .action(DUMMY_CONTRACT_ADDR, getDummyHex(0))
       .action(DUMMY_CONTRACT_ADDR, getDummyHex(1))
@@ -257,7 +273,6 @@ describe("OperationRequestBuilder", () => {
       .refundAsset(stablescam)
       .refundAsset(plutocracy)
       .refundAddr(refundAddr)
-      .chainId(1n)
       .deadline(2n)
       .build();
 
@@ -296,10 +311,14 @@ describe("OperationRequestBuilder", () => {
       refundAddr: refundAddr,
       actions,
       chainId: 1n,
+      tellerContract: DUMMY_CONTRACT_ADDR,
       deadline: 2n,
     };
 
-    const builder = new OperationRequestBuilder();
+    const builder = new OperationRequestBuilder({
+      chainId: 1n,
+      tellerContract: DUMMY_CONTRACT_ADDR,
+    });
     const opRequest = builder
       .action(DUMMY_CONTRACT_ADDR, getDummyHex(0))
       .action(DUMMY_CONTRACT_ADDR, getDummyHex(1))
@@ -310,7 +329,6 @@ describe("OperationRequestBuilder", () => {
       .unwrap(shitcoin, 100n)
       .unwrap(ponzi, 100n)
       .refundAddr(refundAddr)
-      .chainId(1n)
       .deadline(2n)
       .build();